@@ -1,33 +1,36 @@
 [project]
 name = "bentley_ottmann"
+description = "Searching line segments & polygon edges intersections."
 authors = [{ name = "Azat Ibrakov", email = "azatibrakov@gmail.com" }]
 readme = "README.md"
 license = { file = "LICENSE" }
 classifiers = [
     "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: Implementation :: CPython",
-    "Programming Language :: Python :: Implementation :: PyPy",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
     "Programming Language :: Python :: Implementation :: CPython",
     "Programming Language :: Python :: Implementation :: PyPy",
 ]
-requires-python = ">=3.7"
+requires-python = ">=3.10"
 dependencies = [
-    "dendroid>=1.6.1,<2.0",
-    "ground>=9.0.0,<10.0",
+    "dendroid>=2.1.0,<3.0",
+    "ground>=11.0.1,<12.0",
     "prioq>=0.6.0,<1.0",
     "reprit>=0.9.0,<1.0"
 ]
 dynamic = ["version"]
 
 [project.optional-dependencies]
+docs = [
+    "Sphinx>=7.2.6,<9.0",
+    "sphinx-rtd-theme>=2.0.0,<3.0"
+]
 tests = [
-    "hypothesis>=6.70.2,<7.0",
-    "pytest>=7.2.2,<8.0"
+    "hypothesis>=6.148.2,<7.0",
+    "pytest>=9.0.1,<10.0",
 ]
 
 [build-system]
