@@ -1,36 +1,40 @@
 import os
-import sys
 import time
+from collections.abc import Callable, Iterator
 from datetime import timedelta
+from typing import Any, cast
 
 import pytest
-from ground.base import (Context,
-                         get_context)
-from hypothesis import (HealthCheck,
-                        settings)
+from ground.context import Context, get_context
+from hypothesis import HealthCheck, settings
 
-is_pypy = sys.implementation.name == 'pypy'
-on_ci = bool(os.getenv('CI', False))
-max_examples = (-(-settings.default.max_examples // 4)
-                if is_pypy and on_ci
-                else settings.default.max_examples)
-settings.register_profile('default',
-                          deadline=None,
-                          max_examples=max_examples,
-                          suppress_health_check=[HealthCheck.too_slow])
+on_ci = bool(os.getenv('CI'))
+max_examples = settings().max_examples
+settings.register_profile(
+    'default',
+    deadline=(timedelta(hours=1) / max_examples if on_ci else None),
+    max_examples=max_examples,
+    suppress_health_check=[HealthCheck.too_slow],
+    verbosity=2,
+)
+
+# FIXME:
+#  workaround until https://github.com/pytest-dev/pluggy/issues/191 is fixed
+hookimpl = cast(
+    Callable[..., Callable[[Callable[..., None]], Callable[..., None]]],
+    pytest.hookimpl,
+)
 
 if on_ci:
     time_left = timedelta(hours=1)
 
-
-    @pytest.hookimpl(tryfirst=True)
-    def pytest_runtest_call(item: pytest.Item) -> None:
+    @hookimpl(tryfirst=True)
+    def pytest_runtest_call(item: pytest.Function) -> None:
         set_deadline = settings(deadline=time_left / max_examples)
         item.obj = set_deadline(item.obj)
 
-
-    @pytest.fixture(scope='function', autouse=True)
-    def time_function_call() -> None:
+    @pytest.fixture(autouse=True)
+    def time_function_call() -> Iterator[None]:
         start = time.monotonic()
         try:
             yield
@@ -40,13 +44,14 @@ if on_ci:
             time_left = max(duration, time_left) - duration
 
 
-@pytest.hookimpl(trylast=True)
-def pytest_sessionfinish(session: pytest.Session,
-                         exitstatus: pytest.ExitCode) -> None:
+@hookimpl(trylast=True)
+def pytest_sessionfinish(
+    session: pytest.Session, exitstatus: pytest.ExitCode
+) -> None:
     if exitstatus == pytest.ExitCode.NO_TESTS_COLLECTED:
         session.exitstatus = pytest.ExitCode.OK
 
 
 @pytest.fixture(scope='session')
-def context() -> Context:
+def context() -> Context[Any]:
     return get_context()
