@@ -1,2 +1,4 @@
-from .base import (points_strategies,
-                   segments_strategies)
+from .base import (
+    points_strategies as points_strategies,
+    segments_strategies as segments_strategies,
+)
